restore endpoint-publish feature on server side
