+++ webservices/server-integration/src/main/java/org/jboss/as/webservices/publish/EndpointPublisherImpl.java
@@ -36,24 +36,16 @@ import org.apache.catalina.Wrapper;
 import org.apache.catalina.core.StandardContext;
 import org.apache.catalina.startup.ContextConfig;
 import org.apache.tomcat.InstanceManager;
-import org.jboss.as.controller.PathElement;
-import org.jboss.as.server.deployment.DeploymentUnit;
-import org.jboss.as.server.deployment.SimpleAttachable;
 import org.jboss.as.web.deployment.WebCtxLoader;
 import org.jboss.as.webservices.deployers.deployment.DeploymentAspectsProvider;
 import org.jboss.as.webservices.deployers.deployment.WSDeploymentBuilder;
 import org.jboss.as.webservices.service.ServerConfigService;
 import org.jboss.as.webservices.util.WSAttachmentKeys;
 import org.jboss.as.webservices.util.WSServices;
-import org.jboss.as.webservices.util.WebMetaDataHelper;
-import org.jboss.dmr.ModelNode;
 import org.jboss.metadata.javaee.spec.ParamValueMetaData;
 import org.jboss.metadata.web.jboss.JBossServletMetaData;
-import org.jboss.metadata.web.jboss.JBossServletsMetaData;
 import org.jboss.metadata.web.jboss.JBossWebMetaData;
 import org.jboss.metadata.web.spec.ServletMappingMetaData;
-import org.jboss.msc.service.ServiceName;
-import org.jboss.msc.service.ServiceRegistry;
 import org.jboss.msc.service.ServiceTarget;
 import org.jboss.ws.common.deployment.DeploymentAspectManagerImpl;
 import org.jboss.wsf.spi.classloading.ClassLoaderProvider;
@@ -61,7 +53,6 @@ import org.jboss.wsf.spi.deployment.Deployment;
 import org.jboss.wsf.spi.deployment.DeploymentAspect;
 import org.jboss.wsf.spi.deployment.DeploymentAspectManager;
 import org.jboss.wsf.spi.deployment.Endpoint;
-import org.jboss.wsf.spi.deployment.EndpointType;
 import org.jboss.wsf.spi.deployment.WSFServlet;
 import org.jboss.wsf.spi.publish.Context;
 import org.jboss.wsf.spi.publish.EndpointPublisher;
@@ -236,69 +227,5 @@ public final class EndpointPublisherImpl implements EndpointPublisher {
         @Override
         public void destroyInstance(Object o) throws IllegalAccessException, InvocationTargetException {
         }
-    }
-
-    public static class WSEndpointDeploymentUnit extends SimpleAttachable implements DeploymentUnit {
-
-        private String deploymentName;
-
-        public WSEndpointDeploymentUnit(ClassLoader loader, String context, Map<String,String> urlPatternToClassName) {
-            this.deploymentName = context + ".deployment";
-
-            JBossWebMetaData jbossWebMetaData = new JBossWebMetaData();
-            jbossWebMetaData.setContextRoot(context);
-            for (String urlPattern : urlPatternToClassName.keySet()) {
-                addEndpoint(jbossWebMetaData, urlPatternToClassName.get(urlPattern), urlPattern);
-            }
-            this.putAttachment(WSAttachmentKeys.JBOSSWEB_METADATA_KEY, jbossWebMetaData);
-            this.putAttachment(WSAttachmentKeys.CLASSLOADER_KEY, loader);
-        }
-
-        private void addEndpoint(JBossWebMetaData jbossWebMetaData, String className, String urlPattern) {
-            final JBossServletsMetaData servlets = WebMetaDataHelper.getServlets(jbossWebMetaData);
-            WebMetaDataHelper.newServlet(className, className, servlets);
-            final List<ServletMappingMetaData> servletMappings = WebMetaDataHelper.getServletMappings(jbossWebMetaData);
-            if (urlPattern == null) {
-                urlPattern = "/*";
-            } else {
-                urlPattern = urlPattern.trim();
-                if (!urlPattern.startsWith("/")) {
-                    urlPattern = "/" + urlPattern;
-                }
-            }
-            final List<String> urlPatterns = WebMetaDataHelper.getUrlPatterns(urlPattern);
-            WebMetaDataHelper.newServletMapping(className, urlPatterns, servletMappings);
-        }
-
-        @Override
-        public ServiceName getServiceName() {
-            return ServiceName.JBOSS.append("ws-endpoint-deployment").append(deploymentName);
-        }
-
-        @Override
-        public DeploymentUnit getParent() {
-            return null;
-        }
-
-        @Override
-        public String getName() {
-            return deploymentName;
-        }
-
-        @Override
-        public ServiceRegistry getServiceRegistry() {
-            return WSServices.getContainerRegistry();
-        }
-
-        @Override
-        public ModelNode getDeploymentSubsystemModel(String subsystemName) {
-            throw new RuntimeException("Can't get the deployment submodel from a " + WSEndpointDeploymentUnit.class + " instance");
-        }
-
-        @Override
-        public ModelNode createDeploymentSubModel(String subsystemName, PathElement address) {
-            throw new RuntimeException("Can't create a deployment submodel from a " + WSEndpointDeploymentUnit.class + " instance");
-        }
-
     }
 }
